Переносим рендеоринг страницы из 'main.js' в 'PageController'.
